requirements.txt
elevenlabs
google-adk>=1.3.0,<2.0.0
dotenv
streamlit
ffpyplayer
ffmpeg-python
ffmpeg